# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop API - Swagger Documentation",
        "contact": {
            "email": "simonsiva13@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8902/api/v1",
            "description": "Pet Shop API Server"
        }
    ],
    "paths": {
...
        "/order/create": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create order",
                "operationId": "create_order",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "schema": "CreateRequest",
                                "title": "Create Request",
                                "required": [
                                    "order_status_uuid",
                                    "payment_uuid",
                                    "products",
                                    "address"
                                ],
                                "properties": {
                                    "order_status_uuid": {
                                        "description": "Order status UUID",
                                        "type": "string"
                                    },
                                    "payment_uuid": {
                                        "description": "Payment UUID",
                                        "type": "string"
                                    },
                                    "products": {
                                        "description": "Array of objects with product uuid and quantity",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "uuid": {
                                                    "type": "string"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "address": {
                                        "description": "Billing and Shipping address",
                                        "properties": {
                                            "billing": {
                                                "type": "string"
                                            },
                                            "shipping": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/order/{uuid}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Fetch order",
                "operationId": "get_order",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "query",
                        "description": "uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Edit order",
                "operationId": "edit_order",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "query",
                        "description": "uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "schema": "EditRequest",
                                "title": "Edit Request",
                                "required": [
                                    "order_status_uuid",
                                    "payment_uuid",
                                    "products",
                                    "address"
                                ],
                                "properties": {
                                    "order_status_uuid": {
                                        "description": "Order status UUID",
                                        "type": "string"
                                    },
                                    "payment_uuid": {
                                        "description": "Payment UUID",
                                        "type": "string"
                                    },
                                    "products": {
                                        "description": "Array of objects with product uuid and quantity",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "uuid": {
                                                    "type": "string"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "address": {
                                        "description": "Billing and Shipping address",
                                        "properties": {
                                            "billing": {
                                                "type": "string"
                                            },
                                            "shipping": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete order",
                "operationId": "delete_order",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "query",
                        "description": "uuid",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List all orders",
                "operationId": "orders",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "sortBy",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "description": "desc",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}